Fitness
The genetic algorithm is now able to evaluate a graph.
When the graph is set in the a* class, a path can be store to avoid useless
future calculation. A path can be retrieve if it has already been
calculated.
Add bunch of methodes (mainly path calculation oriented) in Geometry.