The genetic algorithme is nearly finished.
Need to review the formula... I am thinking to add average distance
between all vertices as parameter (the more they are spread, the better).
I also add a quick sort algorithm (I have picked it up on Wikipedia).
The Bench.html has now a button that will process a few generations of the
graph (the best one is diplayed on screen) A* still available.